Add files via upload
This is the model + some extra information. We could remove a lot of it if it's too much (especially the comments and displays). Tried to answer as many questions as I could. Used LLMs for the graphs so it could be inaccurate tbh. used yumyumyum's code, just modified it for my purposes when it came to the stuff in the beginning.